[improvement](nereids) Support aggregate functions without from clause
