+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/RuleType.java
@@ -71,6 +71,7 @@ public enum RuleType {
     RESOLVE_PROJECT_ALIAS(RuleTypeClass.REWRITE),
     RESOLVE_AGGREGATE_ALIAS(RuleTypeClass.REWRITE),
     PROJECT_TO_GLOBAL_AGGREGATE(RuleTypeClass.REWRITE),
+    ONE_ROW_RELATION_EXTRACT_AGGREGATE(RuleTypeClass.REWRITE),
     PROJECT_WITH_DISTINCT_TO_AGGREGATE(RuleTypeClass.REWRITE),
     AVG_DISTINCT_TO_SUM_DIV_COUNT(RuleTypeClass.REWRITE),
     ANALYZE_CTE(RuleTypeClass.REWRITE),
