+++ regression-test/suites/nereids_p0/select_no_from/sql/projectAggFuncs.sql
@@ -1,4 +1,17 @@
-/*
 -- database: presto; groups: no_from
-SELECT COUNT(10), MAX(50), MIN(90.0)
-*/
+SELECT 1,
+       'a',
+       COUNT(),
+       SUM(1) + 1,
+       AVG(2) / COUNT(),
+       MAX(3),
+       MIN(4),
+       RANK() OVER() AS w_rank,
+        DENSE_RANK() OVER() AS w_dense_rank,
+        ROW_NUMBER() OVER() AS w_row_number,
+        SUM(5) OVER() AS w_sum,
+        AVG(6) OVER() AS w_avg,
+        COUNT() OVER() AS w_count,
+        MAX(7) OVER() AS w_max,
+        MIN(8) OVER() AS w_min;
+
